+++ lib/stealth/core_ext/string.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+class String
+
+  # Removes blank padding and double+single quotes
+  def normalize
+    self.upcase.gsub(/\"|\'/, '').strip
+  end
+
+end